LPS-74457 Logic that only applies to one check should be kept in that class
